backend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
Django==4.2.6
django-autoslug==1.9.9
django-cleanup==8.0.0
django-colorfield==0.10.1
django-cors-headers==4.3.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
Pillow==10.1.0
pip==23.3.1
psycopg2-binary==2.9.9
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.6.0
